requirements.txt
# DistrictWatch Dependencies
playwright==1.41.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==5.1.0
